chore(@clayui/autocomplete): update Autocomplete component stories
